backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pyodbc==5.2.0
aiosqlite==0.19.0

# Supabase & Database
supabase==2.28.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.11.0
cryptography==41.0.7

# Azure
azure-storage-blob==12.19.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0

# AI/LLM
openai==1.3.8
anthropic==0.7.8

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx[http2]==0.28.1

# Development & Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
